requirements.txt
boto3==1.9.130
botocore==1.12.130
docutils==0.14
jmespath==0.9.4
pynput==1.4.2
python-dateutil==2.8.0
python-xlib==0.25
PyYAML==5.1
s3transfer==0.2.0
six==1.12.0
tqdm==4.31.1
urllib3==1.24.1
